+++ src/game/characters/index.js
@@ -1,0 +1,7 @@
+import { Player } from "./player";
+
+const CHARACTERS = {
+    player: Player
+};
+
+export { CHARACTERS };