+++ src/game/levels/index.js
@@ -1,13 +1,7 @@
 import level_1 from "./level_1";
 
-const levels = [
-    level_1
-];
-
-const LEVELS = {};
-
-levels.forEach((level) => {
-    LEVELS[level.slug] = level;
-});
+const LEVELS = {
+    level_1: level_1
+};
 
 export { LEVELS };