+++ src/game/objects/index.js
@@ -1,0 +1,7 @@
+import { Ground } from "./ground";
+
+const OBJECTS = {
+    ground: Ground
+};
+
+export { OBJECTS };